UGENE-7464. "Select manually" leads to the incorrectly selected region.
Code style changes by review.

+++ src/plugins/pcr_primer_design_for_dna_assembly/src/PCRPrimerDesignForDNAAssemblyOPWidget.cpp
@@ -275,8 +275,7 @@ void PCRPrimerDesignForDNAAssemblyOPWidget::sl_selectManually() {
     int startRegionValue = sbStartRegion->value();
     int endRegionValue = sbEndRegion->value();
     if (startRegionValue > 0) {
-        sequenceSelection->setSelectedRegions({U2Region(static_cast<qint64>(startRegionValue) - 1,
-                                                        static_cast<qint64>(endRegionValue) - startRegionValue + 1)});
+        sequenceSelection->setSelectedRegions({U2Region(startRegionValue - 1, endRegionValue - startRegionValue + 1)});
     }
     updateRegionConnection = connect(sequenceSelection, &DNASequenceSelection::si_onSelectionChanged, this, &PCRPrimerDesignForDNAAssemblyOPWidget::sl_updateRegion);
 }
